newscrab/requirements.txt
blinker==1.8.2
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
ConfigArgParse==1.7
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Login==0.6.3
gevent==24.2.1
geventhttpclient==2.3.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
locust==2.31.8
MarkupSafe==2.1.5
msgpack==1.1.0
psutil==6.0.0
pycparser==2.22
python-dotenv==1.0.1
pywin32==307
pyzmq==26.2.0
requests==2.32.3
setuptools==75.1.0
urllib3==2.2.3
Werkzeug==3.0.4
zope.event==5.0
zope.interface==7.0.3
